Start of pattern for after_compile schema checking
