+++ test/specification/introspection/general/naming_conventions_test.exs
@@ -3,50 +3,29 @@ defmodule Absinthe.Specification.Introspection.General.NamingConventionsTest do
 
   describe "custom types" do
 
-    defmodule PrefixSchema do
-
-      use Absinthe.Schema
-      alias Absinthe.Type
-
-      def query do
-        %Type.Object{
-          fields: fields(
-            __mything: [
-              type: :thing,
-              args: args(
-                __myarg: [type: :integer]
-              ),
-              resolve: fn
-                _, _ ->
-                  {:ok, %{name: "Test"}}
-              end
-            ]
-          )
-        }
-      end
-
-      @absinthe :type
-      def thing do
-        %Type.Object{
-          name: "__MyThing",
-          fields: fields(
-            name: [type: :string]
-          )
-        }
-      end
-
-    end
 
     @standard "Must not define any types, fields, arguments, or any other type system artifact with two leading underscores."
 
     it "cannot start with the `__` prefix" do
-      errs = [
-        "Field `__mything': #{@standard}",
-        "Argument `__myarg': #{@standard}",
-        "Object `__MyThing': #{@standard}"
-      ] |> Enum.sort
-      assert {:error, found_errs} = Absinthe.Schema.verify(PrefixSchema)
-      assert Enum.sort(found_errs) == errs
+      err = assert_raise Absinthe.Schema.Error, fn ->
+        Code.require_file("test/support/specification/introspection/general/prefix_schema.exs")
+      end
+      assert Enum.find(err.problems, fn
+        problem ->
+          match?(%{name: :res_type_name, data: "__MyThing"}, problem)
+      end)
+      assert Enum.find(err.problems, fn
+        problem ->
+          match?(%{name: :res_type_ident, data: :__mything}, problem)
+      end)
+      assert Enum.find(err.problems, fn
+        problem ->
+          match?(%{name: :res_field_name, data: "__mything"}, problem)
+      end)
+      assert Enum.find(err.problems, fn
+        problem ->
+          match?(%{name: :res_arg_name, data: "__mything"}, problem)
+      end)
     end
 
   end
